[dtrans] Recognize left shift as an alloc size multiplier
This change removes the optional parameter indicating whether the check
for an input value as a multiple of size should accept left shift
operations as a multiplier. Previously this was disabled in most case
because the dtrans optimization passes were not able to update the size
in such a case.

Change-Id: I04c29bdef20b8c31667c3a44f9c9450e7ee122ed
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/153347
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Old/split llvm commit ID: 84b3b6edd8515f54408e7238ec1f93db6fd0fd71