[CMPLRLLVM-348] Fix to ArgumentPromotion to avoid debug information of arguments pointing to wrong function
This fixes a problem where debug information becomes invalid following
argument promotion. In the normal case, the debug information about
promoted arguments will be deleted when the original function declaration
is deleted (argument promotion always creates a new function when it
makes changes, and replaces the original function with it), however
there are cases where the original function cannot be deleted, resulting
in the incorrect debug information residing in the new function. This
change immediately updates the debug information for the promoted arguments
when the arguments are updated for the new function, rather than waiting
for the original function to be deallocated.

Change-Id: Ieabc2114ef070efd37c6b06453439e01a0429372
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/154225
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split llvm commit ID: 6a8f28554cadfb76e4348201268606c31c85d864