[dtrans] Skip allocation size checks for ptr-to-ptr types
This change skips the check for allocation size as a multiple of type
size when the type being allocated is a pointer of some kind (meaning
the allocation function returns a pointer-to-pointer value.

Although we may want to shorten the representation of pointers in some
cases, the size of a pointer is well known so we can do this without
locating the size as a constant in the IR.

The change avoids setting "bad alloc size" on the base structure when we
can't find the multiplier used to allocate a buffer of pointers to that
structure.

Change-Id: I9baadf9c21fd3a1af130b46388f753df2254182d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/153346
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Old/split llvm commit ID: 2e89da4660f47d086405ab092edab6e14d0c73f1