Add basic late-outlining support for combined OpenMP directives.
Restructured the code to generate code based on an OMPDirectiveKind
that can be different from the OMPDirectiveKind of the current directive.
This allows processing a combined directive in pieces.

Also did some cleanup to move this closer to what will be needed to
upstream it:

  * Changed routine names with Intel in the name to LateOutline instead.

  * Moved the late-outlining CodeGenFunction members to
    CGIntelStmtOpenMP.cpp.  The main loop processing function is dependent
    on some classes defined in CGStmtOpenMP.cpp so must stay for now.

  * Removed unneeded customization of EmitOMPInnerLoop.

  * Removed \brief commands in Doxygen comments.

Change-Id: Ic32655cab0a11e9a71cdc04bed4f169cc731d2ef
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/154657
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Tested-by: Rice, Michael P <michael.p.rice@intel.com>
Old/split clang commit ID: 02c42b1e690ee09af117f3ace37aecf3875f136f