[dtrans] Handle vtable bitcasts in DTransAnalysis
In preparation for a virtual function call, a pointer to a class object
will often be bitcast as a pointer to the vtable but using the signature
of the function to be called rather than the i32 (...)** form used to
represent the vtable in the class type. This patch recognizes this
secondary form.

CMPLRLLVM-247

Change-Id: I68d1c815807f06f2273deb77a907585e1d4b2c1c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/156483
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split llvm commit ID: 138b344e492eed0b6951fd6a68c855be5dc63313