[dtrans] Handle bitcast access of element zero pointers
If element zero of a structure (outer) is a pointer to another structure (middle) and the pointed-to structure has an instance of a third structure (inner) at element zero, a pointer to outer can be bitcast as a pointer-to-pointer to inner. We weren't recognizing that.

Change-Id: I0b7047f96d00c93f9b4b120e741ba8e4115faa44
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/157294
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split llvm commit ID: f2d0f76b418a029f267a41ffd47a4a884d9e635a