VPO: Provide the hasTargetDirective() util to detect OMP target constructs
In the target compilation for

  void bar() {...}
  void foobar() {...}
  void foo() {
    foobar();
    #pragma omp target
       bar();
  }

the Clang FE filters out foobar(), sending foo() and bar() for offload
codegen in the backend. After VPO outlines the target region in foo(),
it needs to discard foo() but keep bar() in the target code.
To distinguish foo() from bar(), we use this util to see that foo()
contains an OpenMP target region and therefore must be discarded after
outlining is performed.

Change-Id: Ia746b0e45248eb49fa5b6b5a2913bfe3fac28dde
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/157212
Reviewed-by: Lin, Jin <jin.lin@intel.com>
Tested-by: Su, Ernesto <ernesto.su@intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split llvm commit ID: 39eb33f9659484adf4d7674dddb4db1755241f80