[VPlan] [CMPLRLLVM-7328] VPlan vectorizer chooses VF=1 even when #pragma omp simd is specified.
This change fixes an issue where '#pragma omp simd' is not treated as it
should be. We end up using the Profitability check even when it should be
ignored and the loop should be vectorized regardless of its
profitability.

Change-Id: Icfadae897986d792ca07a865ca3ec7601e1495d2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/158033
Reviewed-by: Pushkar Ratnalikar <pushkar.v.ratnalikar@intel.com>
Tested-by: Pushkar Ratnalikar <pushkar.v.ratnalikar@intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split llvm commit ID: 9fe3f0563d5701a4cb70b58ef2719a230244911d