CMPLRLLVM-7142: Replace PHINodes with one incoming value into load/store instructions (XMAIN)
Some BasicBlocks are pruned during inlining. If a BasicBlock is
detached from the predecessor and it have a PHINode with two
incoming values, then the PHINode will be simplified. Now consider
if the "surviving" incoming value is a User of the same PHINode.
The problem is that propagating the "surviving"  value into itself
can destroy the SSA form. This change set is for analyzing if
this propagation could happen, and replacing the PHINode with a
combination of load and stores. The simplification will basically
take the "surviving" incoming value and store it into a temporary
value. This temporary value will be loaded and used to replace
all the uses of the current PHINode.

Change-Id: I7f927087d5c11c6f68caf74b8dd43dee8875b7ee
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/158222
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Old/split llvm commit ID: f5ee5f6745f5c67cb6edb79f46fb957b7135d530