[DTrans] Fix ResolveTypes pass checking for function bitcasts to avoid assertion
When testing compatible types for a type to remap to, resolve types could hit
an assertion if a function bitcast involved a type that was not a member of
same compatibility set. This adds a check that the potential type is compatible.

Change-Id: I95896261b03caa2d2f5bf0f4bd54880515402f09
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/159905
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split llvm commit ID: 6391adbf46ca88f6cb02e6718e647cfcbeb08a09