[VPlan] Optimizations in VPlan HIR decomposer for semi-phi VPInstructions.
This patch updates the HIR decomposer to optimize the creation of
empty semi-phi nodes and updates their placement policy by moving
these instructions to the top of the parent VPBB. The decomposer
now uses a DenseMap to track the created empty semi-phi nodes and
is optimized to create only one unique semi-phi for a given Symbase
within a VPBasicBlock. This is the first patch in the series for the
task of migrating from semi-phi to VPPhiNode in VPlan HIR decomposer.

JIRA : CMPLRLLVM-7243

Change-Id: I781e7cc613ecffac79b2b14596b524bc7ac4b53b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/160769
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Old/split llvm commit ID: 96500965323f534e1d8c86e6317076b398098fa3