[VPlan] Update incoming blocks of VPPHINode when a block is split.
When a VPBasicBlock is split to introduce a new basic block then the
incoming block information of VPPHINodes in the successor blocks must
be updated to use the new generated block. This patch adds support to
perform this update in VPBlockUtils::splitBlock. This is needed to
maintain correctness of VPPHINode i.e. all incoming blocks of a PHI
node must strictly be predecessors of the parent VPBB.

Change-Id: I818658bab772bbc6d4affb1a3c343a3a4b433fcc
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/161654
Reviewed-by: Caballero, Diego <diego.caballero@intel.com>
Tested-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split llvm commit ID: 1f391c98fff31ca5dce6b7d36a8246906130150c