cmplrllvm-7400: accept 0 for #pragma speculated_iterations.
This is allow 0 to be accept for the clause.  One way to fix this is to
disable the clang check and added in sema in our code.

This is from https://git-amr-2.devtools.intel.com/gerrit/#/c/160344/

The alloy run does not working on that.  I am running alloy by hand now.

Change-Id: I6f3d55fb1f257de7d07fb018ad685ab1d32653fe
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/161664
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Tested-by: Yu, Jennifer <jennifer.yu@intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split clang commit ID: 263d935e1096803663548e9f0c08ea496a7c5bf9