[InstCombine] Fold (neg (and (lshr X, C), 1)) -> (ashr (shl X, (BitWidth - C - 1)), (BitWidth - 1))
This pattern corresponds to trying to fill a register with the value of a single bit.

If the bit being tested in X happens to be 7 or 15. The X86 backend can use a movsx to fill the bits above it. Then use a sra to fill in the lower bits.

This gives a 0.66 improvement to 557.xz and a 0.5 improvement to 505.gcc.

Change-Id: Id57c240fb2aa26d431047d9053831e1bd225f7da
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/162678
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split llvm commit ID: 8157daf7d34cf900ab79e191d5555d0ba45c7e1d