[VPlan][NFC] Replace VPUnitaryBlobOrIV with VPOperandHIR class hierarchy
This patch replaces the VPUnitaryBlobOrIV POC that was used to hold the
underlying HIR for blobs and IVs with a more appropriate OO class hierarchy.
The functionality is exactly the same.

As a consequence of introducing the class hierarchy, the DenseMap used as
container for HIR VPExternalDefs had to be replaced by a FoldingSet so that
we can compute the hash of a VPExternalDef without actually creating a
VPExternalDef object. In the future, this container will be replaced with
a SmallVector as discussed in design review and VPlan construction algorithm
will be modified accordingly.

Change-Id: I149211b63aca48a3f15834976d964aa109074142
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/164644
Reviewed-by: Caballero, Diego <diego.caballero@intel.com>
Tested-by: Caballero, Diego <diego.caballero@intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Old/split llvm commit ID: e80dd598f469837b7889b1aa09814298334fb364