CMPLRLLVM-7809: Revert patch D55115 from the community pulldown
The community change in D55115 removes the assumption that
functions are at least 4-byte aligned. The issue is that it
prevents constant folding for cases that deal with function
pointers.For now we are going to revert the changes until we
have a concrete solution to this issue.

Change-Id: I7bd08f41d7309eebdce18d9d1f43d5e8d97638b7
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/164952
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split llvm commit ID: e4e0fdc00c43e3a796e6a71e7f32baa35b964891