[CMPLRLLVM-7708] [DTrans] Lazily recompute DTrans analysis when needed for legacy pass manager.
Do not require DTrans analysis to be recomputed after a pass which does not
modify the types is run when using the legacy pass manager. Mark the analysis
wrapper as invalidated, and when the next pass tries to retrieve the analysis
recompute it.

Change-Id: I84dc5860db0bb45c46a307dc45004e04ae76215d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/165403
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Old/split llvm commit ID: f186730a03b23696ec401771445cf1a714919b08