[LoadCoalescing][NFC] Addressing team code-review comments
This change does away with the use of the temporary MapVector 'Loads'
that was used to collect all the loads in the basic-block. We do not
need the temporary storage and can directly call the 'insertIntoBucket'
method.

Change-Id: I413488c4a2cd97709ff0dcc0174e32815b80d500
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/181653
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Old/split llvm commit ID: 04b1bb56f0717b06a8cc6ad06483ac188fcc4867