[X86] Prefer pcmpeq/pcmpgt+pmovmskb+test over pcmp+kortest. Replace not+pmovmskb with pmovmskb+not.
Compares to mask registers and kortests are higher latency than their legacy equivalents that don't use mask registers. Prefering the older sequences seems to improve the performance of some search loops in xz.

We may consider open sourcing this since it only effects Intel CPUs. We should definitely fix the large diffs this caused on KNL in community as those sequences are obviously bad.

The other change is to move a not from the input to pmovmskb to the output since doing the not in the integer domain is cheaper and can possibly be absorbed by a subsequent compare.

Change-Id: I564dd9c6a58591d1c98a67b3915680a9a9517f8e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/166614
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Old/split llvm commit ID: 6dae3cebb67aba6a05b54a4ee15b71e8a7654d15