[NFC] [DTrans] Change handle to DTransAnalysisInfo in DTransOptBase class from reference to pointer
This changes the DTransOptBase class to store the DTransAnalysisInfo as a pointer
rather than as a reference, and adds some checks that it is non-null. This is to
make the parameter become option in preparation for CMPLRLLVM-7709 which will
allow the base class to be used without the DTransAnalysisInfo for the
ResolveTypes pass.

Change-Id: Ia0af78e9c55466486a448339a6fb5de8df5a6d8e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/166616
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Old/split llvm commit ID: 06d0cb9557a857edc3a424343edb97b24bbb3deb