[VPlan] Add VPOperandHIR information to VPInstructions
This patch is the first step to start attaching the DDRef and IVLevel information
to VPInstructions through the VPOperandHIR class hierarchy so that VPValue-based
code generation can easily have access to the underlying HIR that is still valid.
So far, this information is only added for load VPInstructions, which seems to be
the only case needed at this point.

Testing will be done when code generation uses this information.

Change-Id: I8dd24616997c44cdeec9808a7ff1f4f60dbe5071
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/167496
Reviewed-by: Caballero, Diego <diego.caballero@intel.com>
Tested-by: Caballero, Diego <diego.caballero@intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Old/split llvm commit ID: 8fbdb58d093351aa8212aee29d70509105f28536