Have the compiler emit lics with 0 frequency.
Originally, I meant to reserve 0-frequency for we don't know what the LIC
frequency is. However, it is possible for frequency to underflow to 0. (Nest a
few __builtin_expect edges to get this). In practice, if we don't know what the
LIC frequency is, we're not generating LIC groups in the first place, so I'm
allowing 0-frequency to be emitted.

Change-Id: I82362e69528c4ce749c72a449abb37ca2542ad54
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/168127
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Woodworth, Daniel <daniel.woodworth@intel.com>
Old/split llvm commit ID: 67f2f91a9485b4d3794f4d7333ebcbf50564b1a6