Soften the expensive check: stop recursion if there's a known value.
The relevant methods in SCEV aren't public in open-source LLVM. Intel's
extensions do reset the getSCEVValues function (which is the reverse map of
const SCEV* -> Value*) to be public, so we can use this here.

The actual logic for SCEV expansion is in SCEVExpander::FindValueInExprValueMap.
The check there is more complicated, but for all the queries and the cases we
attempt to use it on here, it should suffice to let us find the cases where the
loop bounds involves a value that already exists but SCEV is more knowledgeable
about how that value itself was calculated.

Change-Id: I39ded8fa437a04c6c27f980505700d6e16f1d65f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/169963
Reviewed-by: Woodworth, Daniel <daniel.woodworth@intel.com>
Reviewed-by: Khaldi, Dounia <dounia.khaldi@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Old/split llvm commit ID: d8ca401f90ef9dc76850281c3c94f8b297e0bf14