[VPlan, NFC] - Move inner loop control flow uniformity transformation utility into a header file.
We are in the process of switching to a new predicator implementation and during this process
the two predicators need to coexist for some time. This change moves the inner loop control
flow uniformity transformation into a header file so that it can be included with the two
predicator implementations.

Change-Id: Ie92a9be3077e37961e2d8eaf3fd1224ef50308a7
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/171089
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Masten, Matt <matt.masten@intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Old/split llvm commit ID: f5f6f13b49a1720e1f2bc33d864dc56deb9d1ecb