[VPO] Check validity of HIR SIMD region before vectorizing the loop.
This patch adds a utility to WRNVecLoopNode to check if the HIR SIMD
region associated to the given loop is valid. The SIMD region is valid
only if both BEGIN and END directives belong to the same parent as the
HLLoop. If the region is invalid we bail out of vectorizer with
appropriate warning message in prod/release compiler, and throw an
assertion failure for debug compiler.

JIRA : CMPLRLLVM-8326

Change-Id: I9e7358e0fc676a4304a6f694f1d7ec6995f95962
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/171782
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Reviewed-by: Panchenko, Nikolay <nikolay.panchenko@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Old/split llvm commit ID: da7ab1bd3a8e89d7f5d5a877620daabe2e249b0e