[WinEH] Allocate a unqiue stack slot for xmm regs in funclets
Currently funclets reuse the same stack slots that are used in the parent
function for saving callee-saved xmm registers. If the parent function modifies
a callee-saved xmm register before an excpetion is thrown, the catch handler
will overwrite the original saved value.

This patch allocates a second stack slot to be used in the EH funclets for
saving these same registers. Long term, it would be better to determine actual
CSR use by the funclets and only allocate the extra space when needed.

This patch was submitted for community review:

https://reviews.llvm.org/D57970

A problem was pointed out there that the automatic stack unwinding
when an exception is thrown from inside a handler would result in the
xmm CSRs being restored from the wrong location because the .seh_savexmm
directive that we are producing is relative to the parent frame rather
than the funclet frame. As I commented there, I believe that this is
only going to be a problem in a very few cases and in all cases my
proposed patch is better than the existing behavior.

Change-Id: I5efe3f0f2b16214cade33778820e9e8218caae67
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/171980
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Pengfei Wang <pengfei.wang@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Old/split llvm commit ID: 17904fd9ce210956439104d789e1010b37d1070d