Enable -prefer-vector-width=256 on SKX by default. Limit ArgumentPromotion pass based on min-legal-vector-width/prefer-vector-width
This patch enables prefer-vector-width=256 by default. And makes ArgumentPromotion conservative if the caller and callee will legalize a 512-bit vector differently. This does not currently check if there are any 512-bit or larger vectors being considered for promotion.

I will try to upstream the ArgumentPromotion part to community as it is a community bug right now. But they might ask me to make it less conservative which will be more work and potentially delay this getting this in and pulled down.

Change-Id: I4bfe7c4b950baac00e0d70cd3f8b49dbd36c0763
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/172658
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Tested-by: Topper, Craig <craig.topper@intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Old/split llvm commit ID: e79f9e63fb85a328b738471b9447a927556f5eb2