// INTEL_FEATURE_ISA_BF16 [X86][BFLOAT16] Correct the implementation of 128-bit CVTNEPS2BF16 intrinsic to reflect that the result is only 64-bits and the upper bits are always zero.
Because the upper 64 bits are always zero, we can't use a 8 element select to represent the masking.

This patch changes the definition of the 128-bit instruction to take a VK4 mask isntead. It modifies the intrinsic for this size to take a v4i1 vector.
A new MCVTNEPS2BF16 ISD opcode is added to reflect a masked version of this. New instruction patterns are added to use this node.

Similar changes were made for VCVTPD2PS, VCVTPD2DQ, VCVTQQ2PS in llorg.
// end INTEL_FEATURE_ISA_BF16

Change-Id: I364abe549ded915db580213a24a3cd2055b32eeb
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/172477
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Tested-by: Topper, Craig <craig.topper@intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Old/split llvm commit ID: 645b306bf2aa6eb62bc4ecf7434ebe59665cf8f3