InferAddressSpaces: rewrite OpenCL builtins
When InferAddressSpaces tries to replace a pointer in a flat address
space (generic AS) and reaches a call to opencl builtin, it should
replace it with a call to the corresponding function that accepts a
concrete address space.

Since all builtin functions in OpenCL are represented by Itanium
mangled function declarations, we have to remangle a function before
we can replace it.

Instead of bringing a full-fledged Itanium mangler-demangler, we
implement a poor-man's mangler that is supposed to work only on
selected functions (57 builtin functions).

Change-Id: I809eb8ea69ceb11ec627f0a7d8e6f2edb5714b2b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/178980
Tested-by: Savonichev, Andrew <andrew.savonichev@intel.com>
Reviewed-by: Savonichev, Andrew <andrew.savonichev@intel.com>
Old/split llvm commit ID: e2d0983db4a3046b69d383be5effb64b28be6152