[CMPLRLLVM-8102] [Devirt/DTrans] Mark bitcast calls created during devirtualization with metadata to help DTrans
The devirtualizer has proven the type of object a virtual call is being
made on. In some cases, this results in a bitcast function call to corece
the object pointer to the class type, the calls need to be marked to
help DTrans analysis to know the bitcast is safe.

Change-Id: I7e143500bdb55c45cd22062c4ed5e083f08663aa
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/173268
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Old/split llvm commit ID: d4d88ba266eafea30affac50ff9548f55040d170