[CMPLRLLVM-8102] [DTrans] Add logic for cascading mismatched argument use when needed
The mismatched argument use safety data needs to handled as being
pointer carried for structure members in many cases. However, there
are some cases where it does not need to be, such as if the
structure field type is the same for both the source structure
type and the type being cast to, or if the structure member
is never actaully used by the program. This change performs
the pointer carried cascading for this safety data, with logic
to detect when it is not necessary to propagate the safety bit.

Change-Id: Ibf475d5997d31c460e2a2276a32fa39985e0824c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/173270
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Old/split llvm commit ID: cd1763c38342b0b6013449d633bf5e84371d5fd3