Sync functions should be marked as convergent
Now declaration of synchronization functions like barrier are marked
with 'noduplicate' attribute. This is necessary to be sure that all
work items will execute the same call to the barrier function, i.e.
a call to barrier function should not be control-dependent. But
noduplicate attribute is too strict for this purpose, convergent
attribute should be used instead. Convergent attribute is different
from noduplicate because it allows duplicating function calls if it can
be proved that the duplicated function calls are not made
control-dependent on any additional values, e.g., unrolling a loop
executed by all work items.

Change-Id: If0f9799dabd39f985323c3c325ca65e89cde1fb1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/174914
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Bader, Alexey <alexey.bader@intel.com>