Add support for select <2 x float> to lowering.
LLVM is normally very finicky about trying to get the types of everything happy.
For the register-or-channel-or-literal operands, we're forced to pick a specific
type by the Operand<ValueType> requirement in the instruction definition. While
perusing the source code of tablegen, I found that there is a hole in the type
logic: iPTR (as a single operand) is converted to the type of its other operand,
if it matches a single type. With this, we can define the relevant type to be
iPTR and then later Pat<> constructs are no longer upset that there is no
intersection of {i64} and {v2f32}, for example.

Change-Id: I5f7fc17945ebff740dd0bb9b8f39e1d40dbcccf1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/175817
Reviewed-by: Woodworth, Daniel <daniel.woodworth@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Cranmer, Joshua <joshua.cranmer@intel.com>
Old/split llvm commit ID: b3349134ae3aab285800ea81ac89aacad3381efb