Prohibiting unlimited LocalSize for numOfComputeUnits == 1
Setting the local size to be the global size works well for small applications,
but since the local size was not limited at it can get very large.

When the JIT is activated it tries to allocate the local size on the stack, so
for such size larger than the stack capacity (machine dependent but typically
around 8MB), when the JIT tries to allocate it on the stack, it causes the
program to segfault as soon as there’s a call in the code.

This patch prohibits assignment of too large values for LocalSize,
even if numOfComputeUnits is 1.

Change-Id: Iabb8daf62a35d7211ed64cb9cdaca5e77518583f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/177417
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>