[X86] Disable canonicalizeBitSelect for vXi1 vectors.
This code is trigger an assertion failure in at least 2 cpu2017 tests in debug builds after the last pulldown.

I don't believe the code in this function applies to vXi1 vectors since it tries to create an X86ISD::ANDNP node which we don't support for vXi1 vectors. Prefering an and+xor instead. Thats not to say we don't potentially need this transform for vXi1, but it would need other modifications to work.

I'm going to push for this fix or some other fix in community, but wanted to unblock the spec failures while I work that out.

Change-Id: I20e84d3ed5bdc1120192ca0408263ddbed5022d6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/179074
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Old/split llvm commit ID: 2e6faf81945c8ad671b429834fb625a8157ca9c2