[CSA][OpenMP] Set external linkage for outlined target regions when emitting binary vISA.
This is a workaround for the limitation in the currently implemented in csa_as
assembling technique that involes two partially linked object files. Symbols
that are created for the outlined target regions and that are defined in the
csa part of the object have to be external because otherwise they cannot be
referenced from the x86 part of the object that contains offload entry table.
This workaround shold be removed once a new assembling technique using two
assembly files is implemented.

Change-Id: I68d9d4b47c73008d124daee7ebc126f5cb2a55bf
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/179321
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>
Tested-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>
Reviewed-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>
Old/split clang commit ID: 9c82d611b34fb1c22c3a4541b31388962a6fc909