[X86] Teach FixupBWInsts to promote some arithmetic operations.
The SelectionDAG code that promotes 16-bit ops to 32-bits does a weak
check for the possibility of using a RMW to memory operation. When we
get to instruction selection we do a stronger check which may fail if
is an unrelated store between the load and store for the RMW. This
leaves behind an unpromoted 16-bit op. This generates an extra 0x66
prefix byte in the encoding which wastes space.

I've seen this in real code with INC/DEC and ADD16ri8. I went ahead and
added all the 16 bit with immediate ops both with 8 bit immediates and
16 bit immediates. Though I don't have a test case for the 16 bit
immediates since we end up favoring folding the load over folding the
immediate. The priority is reversed for 8 bit immediates since we don't
have an 8-bit move immediate into GR16/GR32 instruction

Change-Id: I93770ccb3b50f4b6a924c7379068088ec99b01ab
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/176478
Reviewed-by: Luo, Yuanke <yuanke.luo@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Old/split llvm commit ID: 803c2050a9b66f45dfeabcc04d3eaa7d94e39220