[NFC] CMPLRLLVM-8369: Tighten asserts for recursive progression cloning generation.
There are 2 types of recursive progression clones that can generated. For
the case of #1 (like in 548.exchange2 where there is an explicit terminator)
the formula on Intel_CloneUtils.cpp:237-238 ensures that the LocalCount will
be at least 1. For the case #2 (like 544.nab where the period of the SREM
determines the number of clones), we ensure on Intel_CloneUtils.cpp:283 that
there are at least 2 clones generated.

This means we can assume at Intel_IPCloning.cpp:883 the number of clones will
be at least 1, and we will always go through the loop in lines 884-921. This
means that FirstCloneF will always be non-nullptr at line 923.

Change-Id: Id4c15e6f5e940fa8b47026a67ed50c7be6b68104
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/179612
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Tested-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Old/split llvm commit ID: 80cf097a70c322c87b24dc8229dceec3b6947e12