Use vXi32 vector types for the SVML division and remainder library calls in svmlintrin.h so the MapIntrinToIml pass won't try to scalarize them.
The MapIntrinToIml pass was being confused by the types being vXi64, but the function names were using X*2 as the vector length. This resulted
in the functions being scalarized with a bad scalar function name which caused a link failure.

This patch changes the extern declarations to use a vXi32 type so the function name and the vector length in the type will match.

Change-Id: I55e3a223d7ced9e8ee04f79097ece99a58ed837a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/178309
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Old/split clang commit ID: 4a73eafa42651f8786330ed4f41f925603c50719