[VPO] Move genGlobalPrivatizationLaunderIntrin call to vpo-paropt-pass.
The function is used to make the CodeExtractor generate parameters in
the outlined function for globals and constant exprs. This change is to
localize this instead of doing it in prepare pass. With this, the only
renaming happening in the vpo-paropt-prepare pass will be the ones
necessary to prevent "incorrect" CSE transformations on clause operands.
And for those, there are subsequent changes being planned
(CMPLRLLVM-8608).

Eventually, we will switch to a different way to "trick" the CodeExtractor
instead of overloading the launder intrinsic.

Change-Id: I928caa98741a5a6e60bd625d4a94996b120b7b6e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/180586
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Old/split llvm commit ID: 3700d2471b6e3610fb052e7783b7b79d401689b6