[NFC] CMPLRLLVM-8800: Fix to prevent running LLD in parallel during llvm-lit and prevent early exit [LLD]
The LLD linker for ELF and WebAssembly (WASM) can run multithreaded. The
issue is that if LLVM-LIT runs in parallel, then each call to lld can
spawn more threads and it can cause an exhaustion. This change set
basically adds the environment variable INTEL_LLD_IN_TEST to the lit
configuration, then LLD will read it and prevent running multithreaded
in Windows.

Also, the LLD identifies that it doesn't need to call the destructors
and do an early exit. This was causing memory issues. The call to
exitLld was replaced with exitIntelLld. This function will take care of
removing the temporary data from LTO without exiting early. Then each
driver will always call its proper destructors (during testing and
during regular use).

NOTE: This is a temporary fix and we need to follow up with the
community with the issue related to the destructors.

Change-Id: Id6f2635c3116bc07d79ca0230133f782790ab747
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/182314
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Old/split lld commit ID: 3937448e32dcf3168daa1217f5e53af15e6b4720